I have become totally unsure as to where the train/test split should occur for this dataset, and will seek help on Ed. Have moved all training code from modules.py to train.py, and removed the temporary train/test split form dataset.py which was not working whatsoever.
